+++ src/app/routes.py
@@ -1,9 +1,6 @@
-import sqlalchemy as sa
-from flask import Blueprint, request
-from flask_jwt_extended import create_access_token
+from flask import Blueprint, current_app
 
-from . import db
-from .models.user import User
+from .controllers import auth
 
 bp = Blueprint("index", __name__, url_prefix="")
 
@@ -13,39 +10,9 @@ def index():
     return "Hello, world!"
 
 
-@bp.post("/login")
-def login():
-    data = request.json
-    username = data.get("username")
-    password = data.get("password")
-
-    try:
-        user = db.session.execute(
-            db.select(User).filter_by(username=username)
-        ).scalar_one()
-    except sa.exc.NoResultFound:
-        return {"message": f"No user with username {username}."}
-    if user.check_password(password):
-        token = create_access_token(identity=username)
-        return {"message": "Login successful.", "token": token}
-    else:
-        return {"message": "Invalid credentials."}, 401
-
-
-@bp.post("/register")
-def register():
-    data = request.json
-    username = data.get("username")
-    email = data.get("email")
-    password = data.get("password")
-
-    user = User(username=username, email=email)
-    user.set_password(password)
-
-    db.session.add(user)
-    try:
-        db.session.commit()
-    except sa.exc.IntegrityError:
-        return {"message": f"An user with the username '{username}' already exists."}
-
-    return {"message": "User created successfully."}, 201
+def register_blueprints(app):
+    """
+    Register all blueprints here.
+    """
+    app.register_blueprint(bp)
+    app.register_blueprint(auth.bp)
